1-FE: Rename `utils/errors` to `utils/errorUtils`
